net/os-carp-vip-dhcp: split leaf modules into leasekeeper/ (constants, util, wire)
WIP module split (branch module-split): extract the dependency-free leaves --
constants, util (MAC/IP/mask/jitter/atomic-write/clock helpers) and wire
(neutral BootpFrame/ArpFrame, DhcpReply, reply parse/format/build + the static
capture filter) -- into the new leasekeeper/ package with explicit relative
imports. lease_keeper.py imports the names from the package for now; the
remaining components (codec, capture backends, DhcpClient, policy, Keeper)
follow in later commits, then lease_keeper.py becomes a thin entry point.
123 tests green; full lint pass deferred to the final commit of the split.
