+++ net/os-carp-vip-dhcp/src/opnsense/scripts/OPNsense/CarpVipDhcp/leasekeeper/wire.py
@@ -1,0 +1,102 @@
+"""Backend-neutral wire layer: the frame types both capture backends decode
+into, plus the reply parse/format/build helpers and the static capture filter.
+
+These carry no client state; the stateful protocol sequences live in
+dhcpclient. The option (name, value) vocabulary used here is the keeper's own
+and is deliberately scapy-compatible (ScapyCapture relays outbound option lists
+to scapy verbatim).
+"""
+import re
+from collections import namedtuple
+
+from .constants import DHCP_CLIENT_PORT, DHCP_SERVER_PORT, MTYPE_NAMES, PARAM_REQ_LIST
+
+# A parsed DHCP reply, snapshotted from the capture thread.
+# `message` (DHCP option 56) is the server's optional human-readable text, mainly
+# a NAK reason; `subnet_mask` (option 1) is used to follow a cross-subnet renumber;
+# `giaddr` (BOOTP header) is the relay agent, None when the server is directly
+# attached (no relay in path). The trailing fields are defaulted so existing
+# shorter constructions stay valid.
+DhcpReply = namedtuple("DhcpReply", "mtype yiaddr server_id lease t1 t2 router message subnet_mask giaddr",
+                       defaults=(None, None, None))
+
+# A received BOOTP/DHCP frame in backend-neutral shape: both capture backends
+# (scapy packet / raw bytes) decode to this before the keeper sees it. chaddr
+# is raw bytes; options is a list of (name, value) tuples in the keeper's own
+# option vocabulary (the names in codec's _OPT_ENCODERS/_OPT_DECODERS), which
+# _parse_reply reads regardless of backend.
+BootpFrame = namedtuple("BootpFrame", "op xid yiaddr chaddr giaddr options")
+
+# A received ARP frame (the capture filter already narrows ARP to replies, but
+# op still travels so the handler re-checks rather than trusting the filter).
+ArpFrame = namedtuple("ArpFrame", "op psrc pdst")
+
+# Static BPF capture filter: DHCP (broadcast OFFER/ACK) + ARP replies to our nudge
+# (arp[6:2]=2). A boundary, not an optimization -- it keeps everything else (incl. the
+# segment's broadcast who-has flood) out of the Python parser.
+SNIFFER_FILTER = f"(udp and (port {DHCP_SERVER_PORT} or port {DHCP_CLIENT_PORT})) or (arp and arp[6:2] = 2)"
+
+
+def _msg_text(msg):
+    """DHCP option-56 server text (usually a NAK reason) as a sanitized str, or
+    "" when absent. Option 56 may arrive as bytes or str depending on the server.
+
+    The text is attacker-controlled (any host on the segment can race a NAK with
+    a matching xid) and is written to the log, so strip control characters --
+    newlines that would forge log lines and terminal escape sequences -- before
+    it goes anywhere. Printable content is preserved."""
+    if isinstance(msg, bytes):
+        msg = msg.decode(errors="replace")
+    if not msg:
+        return ""
+    return re.sub(r"[\x00-\x1f\x7f]", "?", str(msg)).strip()
+
+
+def _fmt_reply(rx):
+    """One readable line decoding a received first-party DHCP reply. Logged at
+    DEBUG (the keeper's default level), so every reply's fields (type, addresses,
+    timers, gateway, mask, relay, server text) show in the log without a capture."""
+    txt = _msg_text(rx.message)
+    mtype = MTYPE_NAMES.get(rx.mtype, f"type={rx.mtype}")
+    msg = f" msg={txt!r}" if txt else ""
+    return (f"{mtype} yiaddr={rx.yiaddr or '-'} server={rx.server_id or '-'} "
+            f"giaddr={rx.giaddr or 'none'} lease={'-' if rx.lease is None else rx.lease} "
+            f"t1={'-' if rx.t1 is None else rx.t1} t2={'-' if rx.t2 is None else rx.t2} "
+            f"gw={rx.router or '-'} mask={rx.subnet_mask or '-'}{msg}")
+
+
+def _parse_reply(frame):
+    """Snapshot only the handful of DHCP options the keeper acts on from a
+    BootpFrame into a DhcpReply; the rest of the reply's option data --
+    untrusted, from whatever answered on the wire -- is left untouched."""
+    mt = sid = lt = rt = bt = ro = msg = sm = None
+    for o in frame.options:
+        if isinstance(o, tuple):
+            if o[0] == "message-type":
+                mt = o[1]
+            elif o[0] == "server_id":
+                sid = o[1]
+            elif o[0] == "lease_time":
+                lt = o[1]
+            elif o[0] == "renewal_time":
+                rt = o[1]
+            elif o[0] == "rebinding_time":
+                bt = o[1]
+            elif o[0] == "router":
+                ro = o[1]
+            elif o[0] == "message":     # option 56: server's text (e.g. a NAK reason)
+                msg = o[1]
+            elif o[0] == "subnet_mask":  # option 1: to follow a cross-subnet renumber
+                sm = o[1]
+    gi = frame.giaddr   # relay agent; 0.0.0.0 = directly attached
+    if gi in (None, "0.0.0.0", 0):
+        gi = None
+    return DhcpReply(mt, frame.yiaddr, sid, lt, rt, bt, ro, msg, sm, gi)
+
+
+def _dhcp_options(mtype, extra, id_opts):
+    """The DHCP option list for a message: type, our Parameter Request List
+    (so the server returns the mask/router/timers the keeper acts on), the
+    identity options, then the per-message extras."""
+    return ([("message-type", mtype), ("param_req_list", PARAM_REQ_LIST)]
+            + id_opts + extra + ["end"])
